+++ src/pixelator/pna/analysis_engine.py
@@ -105,6 +105,14 @@ class PerComponentTask(Protocol, Generic[T]):
 
     TASK_NAME: typing.ClassVar[str]
 
+    def setup(self) -> None:
+        """Setup the analysis before running on any components."""  # noqa: D401
+        pass
+
+    def teardown(self) -> None:
+        """Teardown the analysis after running on all components."""
+        pass
+
     def set_dataset(self, pxl_file_path: Path):
         """Specify a dataset to enable analysis being run directly from component IDs."""
         ...
@@ -231,6 +239,14 @@ class AnalysisManager:
             pxl_dataset_builder = PNAPixelDataset.from_pxl_files
         self._pxl_dataset_builder = pxl_dataset_builder
 
+    def _setup(self):
+        for analysis in self.analysis_to_run.values():
+            analysis.setup()
+
+    def _teardown(self):
+        for analysis in self.analysis_to_run.values():
+            analysis.teardown()
+
     def _execute_computations_in_parallel(
         self, component_stream: Iterable[Component | str]
     ):
@@ -305,14 +321,20 @@ class AnalysisManager:
         return self._pxl_dataset_builder(pxl_file_target)  # type: ignore
 
     def _execute_on_iterator(self, iterator, pxl_file_target: PxlFile):
-        if self._n_cores is None or self._n_cores > 1:
-            per_component_results = self._execute_computations_in_parallel(iterator)
-        else:
-            per_component_results = self._execute_computations_sequentially(iterator)
-        post_processed_data = self._post_process(per_component_results)
-        pxl_dataset_with_results = self._add_to_pixel_dataset(
-            post_processed_data, pxl_file_target=pxl_file_target
-        )
+        self._setup()
+        try:
+            if self._n_cores is None or self._n_cores > 1:
+                per_component_results = self._execute_computations_in_parallel(iterator)
+            else:
+                per_component_results = self._execute_computations_sequentially(
+                    iterator
+                )
+            post_processed_data = self._post_process(per_component_results)
+            pxl_dataset_with_results = self._add_to_pixel_dataset(
+                post_processed_data, pxl_file_target=pxl_file_target
+            )
+        finally:
+            self._teardown()
         return pxl_dataset_with_results
 
     def execute(